+++ ms-book/src/main/java/br/com/dev/msbook/responses/Cambio.java
@@ -8,14 +8,14 @@ public class Cambio implements Serializable {
     private Long id;
     private String from;
     private String to;
-    private BigDecimal conversionFactor;
-    private BigDecimal convertedValue;
+    private Double conversionFactor;
+    private Double convertedValue;
     private String environment;
 
     public Cambio() {
     }
 
-    public Cambio(Long id, String from, String to, BigDecimal conversionFactor, BigDecimal convertedValue, String environment) {
+    public Cambio(Long id, String from, String to, Double conversionFactor, Double convertedValue, String environment) {
         this.id = id;
         this.from = from;
         this.to = to;
@@ -48,19 +48,19 @@ public class Cambio implements Serializable {
         this.to = to;
     }
 
-    public BigDecimal getConversionFactor() {
+    public Double getConversionFactor() {
         return conversionFactor;
     }
 
-    public void setConversionFactor(BigDecimal conversionFactor) {
+    public void setConversionFactor(Double conversionFactor) {
         this.conversionFactor = conversionFactor;
     }
 
-    public BigDecimal getConvertedValue() {
+    public Double getConvertedValue() {
         return convertedValue;
     }
 
-    public void setConvertedValue(BigDecimal convertedValue) {
+    public void setConvertedValue(Double convertedValue) {
         this.convertedValue = convertedValue;
     }
